[main] feat: note api 작성

+++ src/main/java/uos/capstone/backend/survey/domain/Survey.java
@@ -26,7 +26,7 @@ public class Survey extends BaseEntity {
 	private Place place;
 
 	@Column(nullable = false)
-	private Integer rate;
+	private Double rate;
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name="userId")
